Move schema file
This is part of a larger goal to separate map specific files from UI files
